Implemented SQLAPI

+++ src/main/java/com/p000ison/dev/simpleclans2/commands/clan/rank/RankDeleteCommand.java
@@ -23,7 +23,7 @@ import com.p000ison.dev.simpleclans2.SimpleClans;
 import com.p000ison.dev.simpleclans2.clan.Clan;
 import com.p000ison.dev.simpleclans2.clanplayer.ClanPlayer;
 import com.p000ison.dev.simpleclans2.commands.GenericPlayerCommand;
-import com.p000ison.dev.simpleclans2.database.data.statements.RemoveRankStatement;
+import com.p000ison.dev.simpleclans2.database.statements.RemoveRankStatement;
 import com.p000ison.dev.simpleclans2.language.Language;
 import com.p000ison.dev.simpleclans2.util.chat.ChatBlock;
 import org.bukkit.ChatColor;
@@ -70,7 +70,7 @@ public class RankDeleteCommand extends GenericPlayerCommand {
             return;
         }
 
-        long response = clan.deleteRank(args[0]);
+        int response = clan.deleteRank(args[0]);
 
         if (response == -1) {
             ChatBlock.sendMessage(player, ChatColor.DARK_RED + Language.getTranslation("rank.not.found"));
